Pull latest C code from https://github.com/facebook/zstd/

This is the code as of https://github.com/facebook/zstd/tree/v1.5.0 (facebook/zstd@a488ba1)

Similar to #73 (comment) to pull changes with edits to flatten the directory structure

Benchmark on a Windows (i7-7700k 4.2GHz w/ 32GB) for mr from http://sun.aei.polsl.pl/~sdeor/index.php?page=silesia:
```
benchstat old.txt new.txt
name                   old time/op    new time/op     delta
CtxCompression-8         92.6ms ± 2%     85.4ms ± 2%   -7.68%  (p=0.000 n=9+9)
CtxDecompression-8       9.44ms ± 1%     9.46ms ± 1%     ~     (p=0.315 n=9+10)
StreamCompression-8       105ms ± 3%       99ms ± 2%   -5.55%  (p=0.000 n=8+10)
StreamDecompression-8    13.7ms ± 9%     12.6ms ± 4%   -7.75%  (p=0.022 n=10+9)
Compression-8             109ms ±12%       86ms ± 2%  -21.00%  (p=0.000 n=9+9)
Decompression-8          10.0ms ± 4%      9.5ms ± 2%   -4.24%  (p=0.001 n=10+9)

name                   old speed      new speed       delta
CtxCompression-8        108MB/s ± 2%    117MB/s ± 2%   +8.33%  (p=0.000 n=9+9)
CtxDecompression-8     1.06GB/s ± 1%   1.05GB/s ± 1%     ~     (p=0.286 n=9+10)
StreamCompression-8    94.7MB/s ± 3%  100.3MB/s ± 2%   +5.88%  (p=0.000 n=8+10)
StreamDecompression-8   733MB/s ±10%    791MB/s ± 4%   +7.96%  (p=0.022 n=10+9)
Compression-8          91.7MB/s ±11%  115.6MB/s ± 2%  +26.05%  (p=0.000 n=9+9)
Decompression-8        1.00GB/s ± 4%   1.05GB/s ± 2%   +4.35%  (p=0.001 n=10+9)
```

Looks like overall performance improvements across the board.