@Viq111 2021-09-10
Thanks for the PR! Do you mind double checking what commit you pulled the files from ?
Looking at https://github.com/facebook/zstd/commit/a488ba1 , there are some changes in this PR that are not in 1.5.0 (example: https://github.com/facebook/zstd/blob/v1.5.0/lib/common/bitstream.h#L334-L335) but are in there

I'm hypothesizing that you might have pulled files from the `dev` branch instead of v1.5.0 (but it changes a lot so hard to know what commit you pulled from.

Do you mind pulling the files from https://github.com/facebook/zstd/commit/a488ba1 instead ?