@SirSniper 2021-09-11
That appears to be exactly what happened. I just pulled the latest code from facebook/zstd@a488ba1 and committed to the branch. Reran my performance results 

```
name                   old time/op    new time/op     delta
CtxCompression-8         87.1ms ± 2%     75.9ms ± 2%  -12.80%  (p=0.000 n=10+9)
CtxDecompression-8       9.60ms ± 1%     9.74ms ± 1%   +1.46%  (p=0.001 n=10+8)
StreamCompression-8       102ms ± 2%       87ms ± 2%  -15.02%  (p=0.000 n=9+9)
StreamDecompression-8    13.1ms ± 5%     12.9ms ± 3%     ~     (p=0.400 n=9+10)
Compression-8            88.1ms ± 3%     76.5ms ± 1%  -13.20%  (p=0.000 n=10+8)
Decompression-8          9.77ms ± 3%     9.93ms ± 4%     ~     (p=0.393 n=10+10)

name                   old speed      new speed       delta
CtxCompression-8        115MB/s ± 2%    131MB/s ± 2%  +14.67%  (p=0.000 n=10+9)
CtxDecompression-8     1.04GB/s ± 1%   1.02GB/s ± 1%   -1.44%  (p=0.001 n=10+8)
StreamCompression-8    97.6MB/s ± 2%  114.8MB/s ± 2%  +17.67%  (p=0.000 n=9+9)
StreamDecompression-8   753MB/s ± 9%    771MB/s ± 3%     ~     (p=0.247 n=10+10)
Compression-8           113MB/s ± 3%    130MB/s ± 1%  +15.19%  (p=0.000 n=10+8)
Decompression-8        1.02GB/s ± 3%   1.01GB/s ± 4%     ~     (p=0.393 n=10+10)

name                   old alloc/op   new alloc/op    delta
CtxCompression-8          56.0B ± 0%      56.0B ± 0%     ~     (all equal)
CtxDecompression-8        64.0B ± 0%      64.0B ± 0%     ~     (all equal)
StreamCompression-8      1.34MB ± 0%     1.07MB ± 4%  -19.63%  (p=0.000 n=8+10)
StreamDecompression-8    7.21kB ±24%     5.76kB ±10%     ~     (p=0.064 n=10+8)
Compression-8             48.0B ± 0%      49.8B ± 8%     ~     (p=0.294 n=8+10)
Decompression-8           48.0B ± 0%      48.0B ± 0%     ~     (all equal)

name                   old allocs/op  new allocs/op   delta
CtxCompression-8           3.00 ± 0%       3.00 ± 0%     ~     (all equal)
CtxDecompression-8         3.00 ± 0%       3.00 ± 0%     ~     (all equal)
StreamCompression-8        3.00 ± 0%       3.00 ± 0%     ~     (all equal)
StreamDecompression-8       161 ± 0%        161 ± 0%     ~     (all equal)
Compression-8              2.00 ± 0%       2.00 ± 0%     ~     (all equal)
Decompression-8            2.00 ± 0%       2.00 ± 0%     ~     (all equal)
```